apps/api/requirements.txt
# === CORE ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# === DATABASE ===
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# === AUTH ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# === HTTP CLIENT ===
httpx==0.26.0
requests==2.31.0

# === UTILS ===
pytz==2024.1
python-dateutil==2.8.2
apscheduler==3.10.4

# === DEV ===
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

